third_party/blink/web_tests/platform/generic/fast/js/prototype-chain-caching-with-impure-get-own-property-slot-traps-expected.txt

Tests what happens when you make prototype chain accesses with impure GetOwnPropertySlot traps in the way.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "function"
PASS typeof f() is "object"
PASS typeof f() is "object"
PASS typeof f() is "object"
PASS typeof f() is "object"
PASS typeof f() is "object"
PASS successfullyParsed is true

TEST COMPLETE
